umap_comparisons/pyproject.toml
[project]
name = "umap-comparisons"
version = "0.1.0"
description = "Poetry package to evaluate UMAP and compare the algo's performance to traditional DR methods"
authors = [
    {name = "paquitopg",email = "pacogoze@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "umap-learn (>=0.5.11)",
    "scikit-learn (>=1.8.0)",
    "scipy (>=1.17.0)",
    "pandas (>=2.3.3)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "numpy (<=2.3)"
]


[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
